+++ plugins/traefik/examples/kubernetes/values.yaml
@@ -1,7 +1,7 @@
 # traefik helm values
 additionalArguments:
   - "--pilot.token=XXXXX_YOURTOKEN_XXXXXXXXXXXXXXXX"
-  - "--experimental.plugins.traefik-ondemand-plugin.modulename=github.com/acouvreur/traefik-ondemand-plugin"
+  - "--experimental.plugins.traefik-ondemand-plugin.modulename=github.com/acouvreur/traefik-ondemand-plugin/plugins/traefik"
   - "--experimental.plugins.traefik-ondemand-plugin.version=v1.2.0"
   - "--providers.kubernetesingress.allowEmptyServices=true"
