+++ plugins/traefik/examples/multiple_containers/docker-stack.yml
@@ -7,7 +7,7 @@ services:
       - --api=true
       - --api.insecure=true
       - --pilot.token=$TRAEFIK_PILOT_TOKEN
-      - --experimental.plugins.traefik-ondemand-plugin.moduleName=github.com/acouvreur/traefik-ondemand-plugin
+      - --experimental.plugins.traefik-ondemand-plugin.moduleName=github.com/acouvreur/traefik-ondemand-plugin/plugins/traefik
       - --experimental.plugins.traefik-ondemand-plugin.version=v1.2.0
       - --providers.docker=true
       - --providers.docker.swarmmode=true
