+++ plugins/traefik/go.mod
@@ -1,6 +1,6 @@
-module github.com/acouvreur/traefik-ondemand-plugin
+module github.com/acouvreur/traefik-ondemand-plugin/plugins/traefik
 
-go 1.17
+go 1.18
 
 require github.com/stretchr/testify v1.6.1
