+++ plugins/traefik/pkg/strategy/dynamic_strategy.go
@@ -5,7 +5,7 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/acouvreur/traefik-ondemand-plugin/pkg/pages"
+	"github.com/acouvreur/traefik-ondemand-plugin/plugins/traefik/pkg/pages"
 )
 
 type DynamicStrategy struct {
