+++ plugins/traefik/traefik.yml
@@ -8,7 +8,7 @@ api:
 experimental:
   plugins:
     traefik-ondemand-plugin:
-      moduleName: "github.com/acouvreur/traefik-ondemand-plugin"
+      moduleName: "github.com/acouvreur/traefik-ondemand-plugin/plugins/traefik"
       version: "v0.1.1"
 
 entryPoints:
